backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "todo-app-backend"
version = "0.1.0"
description = "Backend for the Todo Application"
authors = [{name = "Developer", email = "developer@example.com"}]
dependencies = [
    "fastapi>=0.104.1",
    "sqlmodel>=0.0.16",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "uvicorn[standard]>=0.24.0",
    "psycopg2-binary>=2.9.9",
    "better-exceptions>=0.3.3",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "openai>=1.0.0"
]

[tool.setuptools]
packages = ["agents", "mcp_tools", "routes", "repositories", "alembic"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.10.1",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
    "isort>=5.12.0"
]